% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiCE_function.R
\name{DiCE_function}
\alias{DiCE_function}
\title{Differential Centrality-Ensemble Analysis Based on Gene Expression Profiles and Protein-Protein Interaction Network}
\usage{
DiCE_function(data)
}
\arguments{
\item{data}{A list of 2 elements:
\itemize{
\item \code{data} - A data frame of gene expression data with gene symbols as column names and the last column as class labels (Tumor, Normal).
\item \code{topGenes} - A data frame of DEGs with three columns: \code{Gene.symbol}, \code{adj.P.Val}, and \code{logFC}.
}}
}
\value{
A data frame of key genes with their gene names, log fold change (logFC), and ensemble ranking.
}
\description{
This function offering a comprehensive framework for identifying and prioritizing disease-related genes.
}
\examples{
\dontrun{
  # Example data
  data <- list(
    data = data.frame(gene1 = c(1, 2), gene2 = c(3, 4), class = c("Tumor", "Normal")),
    topGenes = data.frame(Gene.symbol = c("gene1", "gene2"), adj.P.Val = c(0.01, 0.02), logFC = c(2, -2))
  )
  KeyGenes <- DiCE_function(data)
}
}
